Create Geography tree on startup Refactor default tree importing
